Update patch set 1
Patch Set 1:

I feel this board_set_tcpc_power_mode function should be a generic function instead of specific to anx74xx. Before tcpm_init is called, all TCPC must be out of reset, not just anx74xx, and we just happen to have a HOOK_INIT in board.c to take TCPC out of reset, so either add it to tcpci.c and others also, or remove it from anx74xx.c

Also if we use this function to take it in and out of low power mode, tcpm_init or more must be called again to initialize the chip. For example, the interrupt polarity need to be set again.

Patch-set: 1