Align EC host command structures
The host command parameter and response buffers are explicitly aligned
by the LPC/SPI/I2C drivers.  But the host command handlers don't know
that, and the structs are all __packed, so the compiler generates
horribly inefficient ARM Cortex-M code to cope with unaligned accesses.

Force the param / response structs to be aligned, and fix the two
handlers that weren't actually aligned (despite one of them having
comments to the contrary).

Also, add a CHROMIUM_EC define that can be used to determine if a file
is being compiled for an EC target.  We need that so that we only force
structure alignment for EC binaries; on the AP side, the buffers may not
be aligned.

BUG=chromium:647727
BRANCH=any branch which is running out of EC binary size
TEST=Boot a selection of platforms.  If there's no unaligned access
	  exceptions, life is good.

Change-Id: I4537d61a75cf087647e24281288392eb85f22eba