Update patch set 12
Patch Set 12: -Code-Review

UART1 RX can only be used to tell if the system is in s0 or not s0. Cr50 should make sure to release the AP usb whenever the system is not in s0 so it can enable sleep. This patch now uses the ap device state to notify HOOK_CHIPSET_SUSPEND instead of HOOK_CHIPSET_SHUTDOWN.

How to detect shutdown is a different issue that needs to be addressed in a followup cl once we decide on the correct signal to use.

Patch-set: 12