npcx: Selectively filter WDT reset in reset_flags
NPCX reboots by triggering the watchdog, which in turn causes
the watchdog reset flag to be set as one of the reset_flags.
However, other software can confuse the presence of the watchdog
reset flag as a panic.

This patch selectively sets watchdog as a reset flag only
if we're not explicitly doing a soft or hard reset or waking
from hibernate.

Patch created by Mulin Chao <mlchao@nuvoton.com>

BUG=chrome-os-partner:56594
BRANCH=none
TEST=panic reset no longer observed

Change-Id: I016b59ffda4f6334cf41e196960edcbb87f6c049
Signed-off-by: David Hendricks <dhendrix@chromium.org>