power: rk3399: Print RTC during some transitions.
In order to help correlate EC logs with those from the kernel, it was
suggested that the EC could periodically print the RTC time.  This
commit prints out the RTC time when transitioning between S5, S3, and
S0.  Additionally, when a chipset reset is requested, the RTC will be
printed as well.

BUG=chrome-os-partner:57731
BRANCH=gru
TEST=Build and flash kevin.  Boot system up and suspend.  Verify that
RTC logs are printed in between the S5S3, S3S5, S3S0, and S0S3.

Change-Id: Idc9a815c3337f720d41d16e0d844b4c1ea6728d8
Signed-off-by: Aseda Aboagye <aaboagye@google.com>