Update patch set 2
Patch Set 2:

> > Actually, I'm wondering if I should be printing these in the
 > "system" channel instead of the "command" channel.
 > 
 > A new channel perhaps?

I don't think it warrants a new channel, but if it's being printed as a result of the chipset task, it should probably use that channel. If it's being printed as a result from the console command, it should probably use the command channel.

Does that make sense? Or am I overthinking it?

Patch-set: 2