Cr50: Lock console by default, with long timeout
This locks the Cr50 console by default, and makes the unlock
process take the full 5 minutes unless the battery is removed.

BUG=chrome-os-partner:57408
BRANCH=none
TEST=make buildall; test on Gru & Reef

Build the image with

  make buildall

or

  make BOARD=cr50

Put it on Gru & Reef, and confirm that the console is locked, by
running "lock" or "help". Try to unlock it with "lock off" and
observe that you have to poke the power button for five minutes
before it unlocks.

Remove the battery and confirm that "lock off" now only requires
a single power button press to take effect.

Build the image with

  CR50_DEV=1 make BOARD=cr50

and now the behavior is back to the way it was before this CL.

Change-Id: I9f2cc67a1dd63f260221a8711a8591070018be3b
Signed-off-by: Bill Richardson <wfrichar@chromium.org>