Update patch set 1
Patch Set 1:

You should also make sure that it does not break servo. I think that we are going to run into the same contention issues we had with the uart tx signals. So we probably need to disconnect the i2cm peripheral from the gpios whenever servo is connected.

Patch-set: 1