s0ix: Exit s0ix on chipset reset
If a chipset reset happens while the system is in S0iX,
system would continue to remain in S0iX eventhough ap is
in S0. This would cause power state transition and wakeup
issues. To address this, system needs to exit from s0ix as
soon as EC detects a chipset reset

BUG=chrome-os-partner:57920
TEST=enter s0ix, apreset from ec console. EC should exit s0ix
BRANCH=firmware-glados-7820.B

Change-Id: I0f24f717f712dcd46e3ddca2a8c86888739f3deb
Signed-off-by: Jenny TC <jenny.tc@intel.com>