Cr50: Add util/sign_cr50.sh script
This is a quick-and-dirty script to re-sign default images. Our
current Makefiles use only one parameter (CR50_DEV=1) to
determine both the build configuration (prod vs dev) AND whether
or not to sign the binary.

With this, you can build the non-dev binaries like so:

  make BOARD=cr50

and THEN re-sign them for testing without changing anything else
with this:

  ./util/sign_cr50.sh

BUG=none
BRANCH=none
TEST=try it. It works.

Signed-off-by: Bill Richardson <wfrichar@chromium.org>
Change-Id: I776ea017224683dfd7d01a6a97e328bf4dc410b9