kefka: Disable wake up from Keyboard in tablet mode
When system is in S3 and in tablet mode, disable key scanning so that the
system does not resume from sleep,when these periperals are used.

BUG=chrome-os-partner:57316
BRANCH=firmware-strago-7287.B
TEST=Manual on kefka. Boot system with lid open in clamshell
mode. Open a shell window and type powerd_dbus_suspend to bring
the system to suspend state. Now fold the system to tablet mode.
Verify that pressing a key does not cause system to wake up.
Signed-off-by: Chris Wang <chriswang@ami.com.tw>

Change-Id: Ied72855585e1a500d50ba8747bfb04f2cc4b6d48