kefka: Disable wake up from Keyboard or trackpad in tablet mode
When system is in S3 and in tablet mode, disable key scanning and
trackpad so that the system does not resume from sleep,when these
periperals are used.

BUG=chrome-os-partner:57316,chrome-os-partner:57317
BRANCH=firmware-strago-7287.B
TEST=Manual on kefka. Boot system with lid open in clamshell
mode. Open a shell window and type powerd_dbus_suspend to bring
the system to suspend state. Now fold the system to tablet mode.
Verify that pressing a key or trackpad does not cause system to wake up.
Signed-off-by: Chris Wang <chriswang@ami.com.tw>

Change-Id: Ied72855585e1a500d50ba8747bfb04f2cc4b6d48