Update patch set 2
Patch Set 2:

As discussed in https://chromium-review.googlesource.com/#/c/379640/, can we just to a symlink from hana to elm, and ifdef the battery-specific parts?

Unless you think there are going to be large divergences in the future. But, in that case, we can also start with the symlink approach, and duplicate the files later if we feel that's required.

Patch-set: 2
Label: Code-Review=0