Cr50: Add embedded_callback() function to TPM2
This defines a function that the TPM2 library may call to provide
progress reports to us. With it, we can insert minimally
intrusive monitors at various places in the TPM2 library.

The only other way we could know what's happening is to try to
parse the input and output of ExecuteCommand(), which is so
complex that it requires machine-generated marshall/unmarshall
wrappers. It's much simpler and more extensible to let the TPM2
library tell us when something we need to know about is
happening.

BUG=chrome-os-partner:57407
BRANCH=none
CQ-DEPEND=CL:394267
TEST=make buildall; test on Gru

Ensure that the embedded_callback() function is invoked when
TPM2_PCR_Read() is successful, and that it in turn calls
system_process_retry_counter() as before.

Change-Id: Idb9da70c4396eb5d7cfa249449b464dffbf4ae6f
Signed-off-by: Bill Richardson <wfrichar@chromium.org>