usb_updater: always reboot cr50 in the end of the update
With the recent modification of the tpm reset processing the only way
to get the cr50 restart is to reset it internally.

Make sure that usb_updater triggers the cr50 reset in the end of the
update.

BRANCH=none
BUG=none

TEST=with the corresponding init script changes the update on reef
     happens as expected.

Change-Id: Ib49b81c4ef6d12d0b877a8a63493cf4d6d5aaeb0
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/394255
Reviewed-by: Mary Ruthven <mruthven@chromium.org>
(cherry picked from commit 16da7484246d00f37e34a93c97b65cdb41b02dbe)