cr50: modify powerbtn command to simulate press and release
This change adds options to the powerbtn console command to press and
release the power button by using 'powerbtn on' and 'powerbtn off'. It
also adds a 'powerbtn reset' console command to disable the rbox
override. After powerbtn on/off are run powerbtn reset must be called to
stop overriding the state of the physical power button. 'powerbtn pulse'
will automatically disable the rbox override once the pulse is complete.

BUG=chrome-os-partner:58123
BRANCH=none
TEST=manual
	'powerbtn pulse' pulses the power button and disables override
	'powerbtn on' simulates a power button press
	'powerbtn off' releases the power button
	'powerbtn reset' disables the rbox override

Change-Id: I52631d30dbae874ba6637f728cb6e435cb626e12
Signed-off-by: Mary Ruthven <mruthven@chromium.org>