Update patch set 8
Patch Set 8:

I removed all of the board specific stuff from usb_i2c, so it can be moved to common and include directories easily.

But then I would need to add a new config option to differentiate between the need to build the common usb_i2c driver or the stm32 specific one. I'd also need to rename usb_i2c.h in chip/stm32

I think it would be easier to switch stm32 to using this at the same time as moving it to common.

Patch-set: 8