npcx: espi: Fix SMI/SCI handling
The recent change to invert polarity on SMI/SCI is causing problems
for eSPI where the BIOS is seeing ESPI_SMI events that do not clear.

Additionally there appears to be an issue with the PCH in how it
handles eSPI SMI, and if the SMI is not still asserted when it goes
to read SMI_STS then it will not see the ESPI_SMI_STS bit set and
will not know what to service.

So to make eSPI work again leave the negative polarity at its default
setting and instead just set the signal low to send the virtual wire
message as it appears to get reset when the host clears ESPI_SMI_STS.

BUG=chrome-os-partner:58666
BRANCH=none
TEST=boot on eve, shut dowen with lid close at dev screen

Change-Id: Ic860482a945116431e3599473701e733c69d6817
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>