Update patch set 1
Patch Set 1:

> I might be missing something here so I'm open to suggestions.  This
 > is what I found worked the best so far on the "eve" board.

Add Jun in this loop. 
Hi Duncan:
We have a CL for SMI/SCI through VW issue in https://chromium-review.googlesource.com/#/c/400161/. The root cause of it is the reading of SMIB/SCIB will be a fixed value if GPIOC6/GPIO76 are selected as GPIOs. Please review it.

Patch-set: 1
Label: Code-Review=0