apollolake: Offload motion sense initialization to motion sense task
Currently chipset resume takes around 200 ms to complete, and the bulk
of this time is consumed for motion sensor initialization. This keeps
the chipset task busy and delays it in responding to ALL_SYS_PGOOD and
asserting PCH_SYS_PWROK. This impacts S3 resume time. Offloading to
motionsense task eliminates this delay.

BUG=chrome-os-partner:56941,chrome-os-partner:59188
BRANCH=none
TEST=On Reef, bring the system to S3 using powerd_dbus_suspend.
Resume the system by typing powerbtn in EC console and using
the timestamps in EC Log, calculate the time it takes for the
first port80 message after LPC Reset# deassertion to appear. Check
that this time is reduced by approximately 200 ms.

Change-Id: Iec1a292c16b64c324e059ac478176e70a5ebaa68
Signed-off-by: Shamile Khan <shamile.khan@intel.com>