Update patch set 2
Patch Set 2:

> do you mean we need to add it as part of motion sense task?

See ALS drvier driver/als_si114x.c
It was used for Pixel C, so I needed to export it to Android.
AFAIK, ALS were used internally by the EC, not exported to the AP, using a simple interface. If we need a retrofit, we may think about reusing motion_sense framework, allowing the EC to export the sensor to Android.

Patch-set: 2