Update patch set 2
Patch Set 2:

> > Hi Shamile:
 > > Even we applied the CL, we still found host didn't response the
 > VW
 > > SCI on eSPI POC of wheatley. Any suggestion is highly
 > appreciated.
 > 
 > I think you are hitting the same thing I was attempting to fix with
 > https://chromium-review.googlesource.com/#/c/400036/
 > 
 > In short: it seems that if the SMI# VW is not asserted when the SMI
 > handler on the AP goes to read SMI_STS it will not see ESPI_SMI_STS
 > bit set.

Yes, it is. It's necessary to know the current level of SCI/SMI during toggling SCIB/SMIB in HIPMIC. For example, reading HIPMIC is fixed to 0x04 if functionality of SMI/SCI are selected to GPIOs. But CLEAR_BIT(HIPMIC, SCIB) also sets SMIB to zero at the same time. Reading VWEVSM is a good way to solve this issue.

Patch-set: 2