Update patch set 3
Patch Set 3:

I have to test what happens if motion sensors are not up by introducing a delay. It is unlikely, at-least with the set of sensors in reef that this will happen. After removing serialization, I find that the motion sensor initialization completes approximately 50 ms after the first post-resume port80 message appears on EC console. One reason for this is that while sensor initialization is taking place, on the SOC side CSE is executed which  takes 139 ms and there are other delays. Also even once coreboot starts, it will take a while before hands off to Linux occurs.

Patch-set: 3