Update Makefile.rules with a showboards target
This prints all the boards that are part of the build.  This will be
used on coreboot's build system for running coverity.  This allows
the builds to keep in sync as platforms are added, and as anything is
added to the "skip_boards" list.

for board in $(make showboards); do
   make proj-$board CROSS_COMPILE="arm-eabi-"
done

BUG=none
BRANCH=none
TEST=make showboards shows boards, and excludes skip_boards list

Change-Id: I34d8bc95cc5c7d05b4f2ec38b16b75f9f0c72c6f
Signed-off-by: Martin Roth <martinroth@chromium.org>