cr50: Init variable before sending to function that uses it
Although the variable "hashing" is not used for PADDING_MODE_NULL or
PADDING_MODE_PKCS1, the functions DCRYPTO_rsa_encrypt and
DCRYPTO_rsa_decrypt use it for other padding types.

Because of this, GCC 5.3 and newer throws warnings when the variable is
passed in to those functions without being initialized.  To fix this,
always initialize the variable, even if it's not going to be used.

This does not increase the size of any ec.*.flat file.

BRANCH=none
BUG=none
TEST=build succeeds under GCC 4.9.2, 5.3 and 6.2

Change-Id: Iafaaaed8b05080f70f2b1c2f0dbf1ee22227fd78
Signed-off-by: Martin Roth <martinroth@chromium.org>