Update patch set 5
Patch Set 5:

Note that chipset_force_shutdown() is called from other places as well:
* Charger (has its own task; ok)
* Console (apshutdown command; ok)
* Thermal (uses HOOK_SECOND; problem!)

The problem is that HOOK_TICK (which resets the watchdog) and HOOK_SECOND (for thermal) use the same task.  So if we sleep it for 10 seconds, then the EC will watchdog-reboot after a second, which will force-shutdown the AP.

When the EC reboots, it'll attempt to power the AP back on to do software sync.  If the over-temperature condition still exists, then thermal_control() will attempt to shut down the system, which will again watchdog-reboot after a second.

A sufficiently hot system will thus reboot once a second, triggering the worst case behavior.

Something like the following may work:

while(time_diff < SHUTDOWN_DELAY) {
  /* If forcing shutdown from a hook, we need to explicitly poke the watchdog */ 
  usleep(250 * MSEC);
   watchdog_reload();
   time_diff += 250 * MSEC;
}

Patch-set: 5