caroline: disable trackpad in tablet mode
The (new) TRACKPAD_INT_DISABLE GPIO is preventing the trackpad interrupt line going
to the AP from triggering events (when we are in tablet mode).

Note: on previous hardware without this modification, this will likely
prevent the trackpad from working.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=glados
BUG=chrome-os-partner:57993
TEST=put the laptop in tablet mode, see that trackpad input is ignored,
go back to regular laptop mode and see that trackpad is working.

Change-Id: Ie9577d93cf058b102f50c109240c5d4bd6af5cb3