cr50: make unlock sequence more user friendly
Unlocking CR50 console is a procedure requiring the user to assert
physical presence (by pressing the 'power' button of the device). It
is supposed to take a significant amount of time too.

Let's not require the user to poke the power button too often - once
in 30 seconds should be enough. Also, to make it easier, print
countdown until the next poke deadline.

BRANCH=none
BUG=none
TEST=verified that unlocking cr50 console still is possible, but much
     lower poking rate is required and the poke deadline is properly
     displayed.

Change-Id: I1ef1d65b66556a4c5a567ad77372c74993e1febe
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>