power_button_x86: avoid chipset force shutdown from off state
This is to avoid system from shutting down asynchronously if the
power button is pressed and held for more than 4 seconds since x86
chipset will force itself to shutdown based on its own hardware
timer on 4 seconds timeout.

The current code has already accounted for the case when Refresh + Power
Button is pressed and held to not shutdown by not continue to pass
through the power button signal after the 200ms pulse is sent. This
change is follownig the same behavior. The lid open execution path from
off state is also similar.

BUG=chrome-os-partner:58513
BRANCH=firmware-terra-7287.154.B
TEST=System does not shut itself off if the power button is pressed and
held from S5. The power button is behaving as expected in S0 or S3.

Change-Id: I52536749ad6f4d1a3fcf6ba4c5b3bb6f953da6a5
Signed-off-by: Kevin K Wong <kevin.k.wong@intel.com>