Update patch set 1
Patch Set 1:

> We'll need to put this behind a CONFIG option to land on glados branch / TOT, but it's fine for terra branch.

In the case that system is OFF, if the user press and hold Refresh + Pwrbtn vs PwrBtn alone, the existing code will ignore the power button hold since it took the PWRBTN_STATE_BOOT_KB_RESET path, so this change is giving consistence behavior between holding PwrBtn alone and Refresh  Pwrbtn. Add CONFIG flag may make sense to glados branch if we don't want to change the behavior on shipping system, but does TOT need a new flag if this is ported?

Patch-set: 1