Update patch set 2
Patch Set 2: Code-Review-1

The goal is to replicate the level of interactivity required to remove the back panel screws and the WP screw.  Tapping every 30 seconds is a significantly lower bar.

Note that it would be trivial for someone to modify a USB stick computer to act as both SuzyQ and host.  So all it'd look like is someone plugging in a USB key and tapping power every 30 seconds, and bam, you're owned.

Patch-set: 2
Label: Code-Review=-1