Update patch set 9
Patch Set 9: Code-Review-1

(8 comments)

Mostly looks good, I'm glad to see the skylake and apollolake code coming together.

As you said much of this patch is just reshuffling code. It's the rest that we need to look at closely, even for small stuff like:
- Merging lpc_clear_host_events() into lpc_disable_wake_mask_for_lid_open().
- Adding CHIPSET_POWER_WAIT_SIGNALS.
- Changing the return type of chipset_force_g3().
- Moving certain parts of the power state machine from _power_handle_state() to power_handle_state().

One that's not so small:
- Change handling of S0IX, specifically with regard to sampling PCH_SLP_S0_L versus relying on host events. This seems to be a completely unrelated change that was not present in Skylake nor Apollolake before. I would like to better understand the implications before we merge any patch with this change.

FWIW we can begin to merge much of this patch if it's split into a series of smaller patches. It's perfectly fine to have patches that do nothing but cut+paste or rename stuff. That will also help avoid a huge mess if we need to revert one of the non-trivial changes.

Patch-set: 9
Label: Code-Review=-1