Update patch set 1
Patch Set 1: Code-Review-1

I wasn't really involved in the implementation of virtual battery, but IMHO this is actually the wrong direction to go.  I'd rather virtual battery actually fully implement the interface needed by the kernel and remove the "fallback to the battery" case.  Then the solution is to just make sure that the EC itself is self-consistent and you're done.

I haven't looked at what commands the kernel uses that the EC is missing, so I don't know if that's truly practical, but that was the original intention when I suggested the virtual battery.

Patch-set: 1
Reviewer: Gerrit User 1000107 <1000107@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1