Update patch set 1
Patch Set 1:

I haven't dug completely deeply, but this seems odd to me.  Doesn't the host call host_get_next_event() and doesn't that clear the events?

In general I worry about clearing interrupts like this on the suspend path because if there was a valid wakeup event right at the time we're going to suspend then it will be dropped (AKA user sees screen turn off and tries to hit a key to wake the device back up right away).

Patch-set: 1