Update patch set 9
Patch Set 9:

we sure should land this, Scott, please add a counter of corrected errors and a cli command to report its value and reset it if required.

Also, your idea of keeping a log of most reset occurrences with timestamps is useful, but let's leave it for the next patch as an enhancement.

Patch-set: 9