Update patch set 2
Patch Set 2:

I don't know the history, but it also seems fairly odd to me that the mutex is only there for the "add" case and not there for any other FIFO operations.  Are we certain that no other FIFO operations are called by more than one thread?

To put it another way, it appears that the current mutex only helps if you've got concurrent adders.  Then we rely on the "atomic" operations to make sure that add/remove are safe with respect to each other.  ...and I guess we rely on only a single thread ever doing removes?

That doesn't actually seem true, really.  It appears that keyboard_clear_buffer() can at least be called via two contexts and that grabs no locks at all...

Seems like the whole locking scheme for this FIFO needs to be analyzed...

Patch-set: 2