cr50: disable bus obfuscation
When bus obfuscation is enabled it interferes with cr50 resuming from
sleep. It is currently enabled by the fuses. This change overrides the
fuses to disable bus obfuscation.

In fuses.c I copied all of the FUSE_PROG registers from the hardware
description, and used a script to populate init_fuses based those names.

BUG=b:32649680
BUG=chrome-os-partner:57994
BRANCH=none
TEST=put cr50 to sleep for 30 minutes. Try to wake it up over uart and
verify it doesn't crash

Change-Id: I4d7ba5e97f3d2bd72528f647459109c7646df414
Signed-off-by: Mary Ruthven <mruthven@chromium.org>