Update patch set 1
Patch Set 1:

Seems like a good idea (I would not vouch for the crypto-analysis but I trust Marius),
but you are obfuscating even more that the previous version was using exponent 'F4' (the only hint was the mod_pow_F4 function that you have renamed).
Might worth clarifying that the configuration choice is between exponent 3 and exponent 65537.

Patch-set: 1
Label: Code-Review=0