Cr50: Add TPM commands to get/set the console lock
This adds TPM vendor-specific commands to get the state of the
Cr50 console lock, or to set the lock. There is intentionally no
command to unlock the console.

This just adds the commands to the Cr50. We still need to provide
a nice userspace utility for the AP to use.

BUG=chrome-os-partner:57940
BRANCH=none
TEST=make buildall; load on Reef, test manually

On Reef, I can use the trunks_send tool to send the raw TPM bytes
to invoke these commands:

Get the lock state:

 # trunks_send 80 01 00 00 00 0a 20 00 00 00
  80010000000B0000000000

The last byte 00 indicates that the lock is NOT set, so set it:

 # trunks_send 80 01 00 00 00 0a 20 00 00 01
  80010000000A00000000

Success. On the Cr50 console, I see it take effect:

  [480.080444 The console is locked]

Query the state again:

 # trunks_send 80 01 00 00 00 0a 20 00 00 00
  80010000000B0000000001

and now the last byte 01 indicates that it is set.

Change-Id: I17165dcad667b47a6307fe0ade7284ef5b3005d6
Signed-off-by: Bill Richardson <wfrichar@chromium.org>