Update patch set 1
Patch Set 1:

> This does not increase the size of ANY ec.rX.flat files.

Thanks for this interesting bikeshedding,
if you look at the disassembly of the former version,
GCC was already putting 0 there on the stack when printing the undefined value.
but 0 is not the value we want 
So yes your fix is just as bad as the previous version if that's what you are trying to demonstrate.
Else if you are trying to fix this code,
you can move up 15 lines the block_offset initialization:
block_offset = be32toh(cmd_body->block_base);
and you will see this is saving the 2 bytes of putting the dummy 0 inside this variable.

Patch-set: 1