g: Indicate invalid firmware images in version string
If an inactive firmware slot is marked invalid, the version
string should say so. This prefaces an invalid firmware version
with "XX".

BUG=none
BRANCH=none
TEST=make buildall; manual check on running system

Start with all RO/RW A/B images valid:

  > version
  Chip:    g cr50 B2
  Board:   0
  RO_A:  * 0.0.10/2ccaf5c7
  RO_B:    0.0.9/60eccbb7
  RW_A:  * 0.0.9/DEV/cr50_v1.1.5590-d971238
  RW_B:    0.0.9/DEV/cr50_v1.1.5589-3e58bda
  Build:   0.0.9/DEV/cr50_v1.1.5590-d971238
           tpm2:v0.0.279-07ec984
           cryptoc:v0.0.6-7c350b9
           2016-11-11 17:44:02 wfrichar@wintermute.mtv.corp.google.com
  >

In this example, I'm running from RO_A and RW_A, so we'll poke at
the B images. The flashwrite command uses OFFSET, not ADDRESS, so
the RO_B and RW_B firmwares are at offset 0x40000 and 0x44000
respectively.

Scribble over the first 4 bytes of the RW_B header, and notice the RW_B
version has changed:

  > flashwrite 0x44000 4
  Writing 4 bytes to 0x44000...
  > version
  Chip:    g cr50 B2
  Board:   0
  RO_A:  * 0.0.10/2ccaf5c7
  RO_B:    0.0.9/60eccbb7
  RW_A:  * 0.0.9/DEV/cr50_v1.1.5590-d971238
  RW_B:    XX0.0.9/DEV/cr50_v1.1.5589-3e58bda
  Build:   0.0.9/DEV/cr50_v1.1.5590-d971238
           tpm2:v0.0.279-07ec984
           cryptoc:v0.0.6-7c350b9
           2016-11-11 17:44:02 wfrichar@wintermute.mtv.corp.google.com
  >

But we can't scribble over RO_B:

  > flashwrite 0x40000 4
  Writing 4 bytes to 0x40000...
  do_flash_op:237 errors 4000 fsh_pe_control 40720004
  do_flash_op:271, retry count 9
  Unknown error
  Usage: flashwrite offset [size]
  >

It's protected by GLOBALSEC flash region 6. We'll have to open a window to
it. Note that the base register uses the ADDRESS, not the OFFSET:

  > rw 0x40090260 0x00080000
  > rw 0x40090264 0x1fff
  > rw 0x400900f8 7

Now it works:

  > flashwrite 0x40000 4
  Writing 4 bytes to 0x40000...
  > version
  Chip:    g cr50 B2
  Board:   0
  RO_A:  * 0.0.10/2ccaf5c7
  RO_B:    XX0.0.9/60eccbb7
  RW_A:  * 0.0.9/DEV/cr50_v1.1.5590-d971238
  RW_B:    XX0.0.9/DEV/cr50_v1.1.5589-3e58bda
  Build:   0.0.9/DEV/cr50_v1.1.5590-d971238
           tpm2:v0.0.279-07ec984
           cryptoc:v0.0.6-7c350b9
           2016-11-11 17:44:02 wfrichar@wintermute.mtv.corp.google.com
  >

Change-Id: I537a2de157018046f04ed14ff35467c9236d8f1c
Signed-off-by: Bill Richardson <wfrichar@chromium.org>