cr50: add a function to tell between prod and dev keys
To keep things simple it has been decided to use a certain key ID bit
as an indication of a prod versus dev key type.

The key ID value is derived from the key value, and is easy to enforce
(by generating another key in case the value does not match the
required key type).

This information will also be available through the cr50 userspace
tool.

BRANCH=none
BUG=chrome-os-partner:58230
TEST=ran the new code on prod and dev devices; On prod:
   > sysinfo
   Reset flags: 0x00000800 (hard)
   Chip:        g cr50 B2
   RO keyid:    0x3716ee6b(dev)
   RW keyid:    0xb93d6539(dev)
   DEV_ID:      0x015bb2a9 0x04656742

    on dev:
   > sysinfo
   Reset flags: 0x00000800 (hard)
   Chip:        g cr50 B2
   RO keyid:    0xaa66150f(prod)
   RW keyid:    0xde88588d(prod)
   DEV_ID:      0x0180305d 0x04656742

Change-Id: I37b12e02eb62f4f86bb58468ee301e77446451da
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>