chip/stm32/usb_hid_keyboard: Fix set_keyboard_report race
We always want set_keyboard_report to send the freshest possible
data. For this purpose, we use double-buffering on the USB
endpoint.

To make sure there is no race between the address change and
the data being transmitted, we set an atomic, asking the
interrupt handler to reload the next buffer. This will send
an extra (identical) report, but those are ignored by the
HID subsystem.

BRANCH=none
BUG=chrome-os-partner:59083
TEST=make buildall -j
TEST=make BOARD=hammer -j && util/flash_ec --board=hammer

Change-Id: I9d14541b8b05017c1d5051b9a315db381a89dcea